DETAILS - Spec functionality almost all complete
Complete

- 4 sections with required inputs:
Appearance, Transactions, Server and SMTP Integration

Excepting

- Validation : Not clear how best to do validation of nested models.
- Local settings : Splitting some of SMTP integration's settings into local settings
and some into the settings api store. It is unclear why and the best
way to do this.
- API format : utils/settingsAPIAdaptor.js contains an explanation of the
difference between the API format and the model format and especially how
3 keys are omitted from the api, being those for Appearance, Transactions
and Server settings. Requesting an API extension to permit storing these keys
is proposed.

Other Issues

- There is still much I do not know about how data flows through this application.
- There was a strange bug discovered where, the code that attempts to facilitate nesting models,
also ends up deleting the keys of nested models. Existing code, like avatarHashes depends on this,
and the best way to really implement nested models in a maintainable architecture is something it works
to get some clarity around. The existing model nesting code appears to be a "partial solution" to the
issue of nesting models. Possible forward paths involve having separate APIs for all models and "denormalizing"
models ( ShippingAddress, avatar ) so there is no nesting. This is the least preferred option. Another
option is to build ( or use ) a complete solution for nesting Models in backbone. One that brings clarity to all things like syncing, change events, and so on.