New CSS to match the designs, with the color of the chat close button. Also makes improvements to the test notice, and prevents the chat from covering the scrollbar.
Note that this does not match the designs when the close button is over a background with no tab bar, in that case the button is the color of the background instead of the primary color. But after reviewing the designs, I can't think of a situation in which that would actually happen, there should always be a tab bar present.