Improve Error Check
In the setModelQuantity method, an error is thrown when cur is not a string, even though cur is only used when changing the quantity of cryptocurrency listings. This works currently because the API returns a metadata.coinType value of "", but when fetching from IPFS that value is missing for non-crypto listings.